+++ R/utils_set_attr.R
@@ -1,4 +1,4 @@
-set_attributes_and_class <- function(data, model, t.title, x.title, y.title, l.title, legend.labels, x.axis.labels, faminfo, x.is.factor, full.data, constant.values = NULL, terms = NULL) {
+set_attributes_and_class <- function(data, model, t.title, x.title, y.title, l.title, legend.labels, x.axis.labels, faminfo, x.is.factor, full.data, constant.values = NULL, terms = NULL, n.trials = NULL) {
   # check correct labels
   if (!is.null(x.axis.labels) && length(x.axis.labels) != length(stats::na.omit(unique(data$x))))
     x.axis.labels <- as.vector(sort(stats::na.omit(unique(data$x))))
@@ -21,6 +21,8 @@ set_attributes_and_class <- function(data, model, t.title, x.title, y.title, l.t
   attr(data, "family") <- faminfo$family
   attr(data, "link") <- faminfo$link.fun
   attr(data, "logistic") <- ifelse(faminfo$is_bin, "1", "0")
+  attr(data, "is.trial") <- ifelse(faminfo$is_trial && inherits(model, "brmsfit"), "1", "0")
+  attr(data, "n.trials") <- n.trials
 
   # and model-function
   attr(data, "fitfun") <- get_model_function(model)